pyproject.toml
[tool.poetry]
name = "hipaah"
version = "0.1.0"
description = "Developer-first HIPAA authorization and access control system"
authors = ["Jeff Park <aventyrlabs@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "hipaah" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pyyaml = "^6.0"
typer = { extras = ["all"], version = "^0.9.0" }
faker = "^24.0.0"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
pre-commit = "^3.6.0"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
